Fixed issue where unsigned ints were being compared with millis(). Now using unsigned long.

+++ MR_Solenoid.cpp
@@ -16,14 +16,14 @@ unsigned short int _SOLENOID_MIN_ON_TIME =1;
 unsigned short int _SOLENOID_MAX_ON_TIME = 255;
 
 //timer to figure out when to turn off solenoid
-unsigned int _SOLENOID_1_LAST_TRIGGER_TIME = 0;
-unsigned int _SOLENOID_2_LAST_TRIGGER_TIME = 0;
-unsigned int _SOLENOID_3_LAST_TRIGGER_TIME = 0;
-unsigned int _SOLENOID_4_LAST_TRIGGER_TIME = 0;
-unsigned int _SOLENOID_5_LAST_TRIGGER_TIME = 0;
-unsigned int _SOLENOID_6_LAST_TRIGGER_TIME = 0;
-unsigned int _SOLENOID_7_LAST_TRIGGER_TIME = 0;
-unsigned int _SOLENOID_8_LAST_TRIGGER_TIME = 0;
+unsigned long _SOLENOID_1_LAST_TRIGGER_TIME = 0;
+unsigned long _SOLENOID_2_LAST_TRIGGER_TIME = 0;
+unsigned long _SOLENOID_3_LAST_TRIGGER_TIME = 0;
+unsigned long _SOLENOID_4_LAST_TRIGGER_TIME = 0;
+unsigned long _SOLENOID_5_LAST_TRIGGER_TIME = 0;
+unsigned long _SOLENOID_6_LAST_TRIGGER_TIME = 0;
+unsigned long _SOLENOID_7_LAST_TRIGGER_TIME = 0;
+unsigned long _SOLENOID_8_LAST_TRIGGER_TIME = 0;
 
 //states used for diplaying status 
 unsigned int _SOLENOID_1_STATE = 0;//assume 0 to be idle (returned and resting);
